Correzioni di vari errori e introduzione del bottone di avvio per l'IF-16
